fix(src) fix some potential problem in lu_hash_func and table init func

+++ luhash.c
@@ -10,7 +10,7 @@
  *
  * @author [hesphoros]
  * @contact [hesphoros@gmail.com]
- * @date 2025-1-1
+ * @date 2025-1-15
  * @version 1.0
  */
 
@@ -39,7 +39,7 @@ static void lu_hash_list_destory(lu_hash_bucket_t* bucket);
 static void lu_hash_rb_tree_destory(lu_hash_bucket_t* bucket);
 
 static lu_rb_tree_node_t* lu_rb_tree_successor(lu_rb_tree_t* tree, lu_rb_tree_node_t* node);
-static int	lu_hash_function(int key, int table_size);
+static int	lu_hash_function(int key, size_t table_size);
 
 static void lu_hash_table_resize(lu_hash_table_t* table);
 
@@ -58,7 +58,7 @@ static void lu_rb_tree_rehash(lu_rb_tree_t* tree, lu_rb_tree_node_t* node, lu_ha
  * @param table_size The size of the hash table (number of buckets).
  * @return The computed hash value, ranging from 0 to table_size - 1.
  */
-static int lu_hash_function(int key, int table_size)
+static int lu_hash_function(int key, size_t table_size)
 {
 	static const double golden_rate_reciprocal = 0.6180339887; // Reciprocal of the golden ratio
